Move modules from `/app`
