--- apps/meteor/ee/app/auditing/client/templates/index.ts
@@ -1,1 +1,0 @@
-import './audit/audit.ts';
