+++ apps/meteor/ee/client/templates/audit/audit.css
