+++ apps/meteor/ee/definition/index.ts
@@ -1,3 +1,2 @@
-import '../app/auditing/definition/methods';
-import '../app/license/definition/methods';
+import './methods';
 import './rest';
