+++ apps/meteor/ee/definition/methods/index.ts
@@ -1,0 +1,2 @@
+import './audit';
+import './license';
