+++ apps/meteor/ee/server/index.ts
@@ -3,7 +3,6 @@ import './startup';
 import '../app/models/server';
 import '../app/license/server/index';
 import '../app/api-enterprise/server/index';
-import '../app/auditing/server/index';
 import '../app/authorization/server/index';
 import '../app/canned-responses/server/index';
 import '../app/livechat-enterprise/server/index';
