+++ apps/meteor/ee/server/startup/index.ts
@@ -1,3 +1,4 @@
+import './audit';
 import './deviceManagement';
 import './engagementDashboard';
 import './seatsCap';
